refactor(macos): extract MacOSComponentPresenter from AppModel (Stage 2)
Move the macOS AppKit presentation out of `AppModel` into a dedicated
`MacOSComponentPresenter`: it owns the presented `contentViewController`,
the window title, and the default/modal presentation (including the
`presentAsModalWindow` + `.RCTContentDidAppear` modal-sizing path).

`AppModel` no longer constructs `NSViewController`s; it holds the shared
`ComponentPickerModel`, the `ReactInstance`, React init/registration, and
bundle-source actions, and delegates component presentation to the
presenter. `RootContentView` observes the presenter for multi-app content
and title. This decouples the model from AppKit and is the enabler for
sharing `AppModel` with iOS in a later stage.

Config-plugin anchor comments are unchanged. macOS-only; iOS is untouched.

Verified locally against react-native-macos 0.81: swiftc -parse (4 config
combos), swiftformat --lint, swiftlint --strict, plutil, and
node --test test/pack.test.mts all pass; full macOS build succeeds and the
app launches with the React menu, component selection, and modal
presentation all working through the presenter.
